requirements.txt
altair==5.2.0
asttokens==2.4.1
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.42
huggingface-hub==0.21.4
idna==3.6
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
packaging==23.2
pandas==2.2.1
parso==0.8.3
Pillow==8.4.0
prompt-toolkit==3.0.43
protobuf==4.25.3
pure-eval==0.2.2
pyarrow==15.0.2
pydeck==0.8.1b0
pygame==2.5.2
Pygments==2.17.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.6.0
sentencepiece==0.2.0
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
streamlit==1.32.2
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.4.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.1
torchvision==0.17.1
tornado==6.4
tqdm==4.62.3
traitlets==5.14.2
transformers==4.39.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
wcwidth==0.2.13
